+++ medusa/server/core.py
@@ -35,6 +35,7 @@ from medusa.server.api.v2.schedule import ScheduleHandler
 from medusa.server.api.v2.search import SearchHandler
 from medusa.server.api.v2.series import SeriesHandler
 from medusa.server.api.v2.series_asset import SeriesAssetHandler
+from medusa.server.api.v2.series_change_indexer import SeriesChangeIndexer
 from medusa.server.api.v2.series_legacy import SeriesLegacyHandler
 from medusa.server.api.v2.series_mass_edit import SeriesMassEdit
 from medusa.server.api.v2.series_mass_operation import SeriesMassOperation
@@ -117,6 +118,9 @@ def get_apiv2_handlers(base):
         SeriesMassEdit.create_app_handler(base),
         # /api/v2/massupdate
         SeriesMassOperation.create_app_handler(base),
+
+        # /api/v2/series/changeindexer
+        SeriesChangeIndexer.create_app_handler(base),
         # /api/v2/series/tvdb1234/operation
         SeriesOperationHandler.create_app_handler(base),
         # /api/v2/series/tvdb1234/asset