+++ medusa/server/web/manage/handler.py
@@ -36,6 +36,14 @@ class Manage(Home, WebRoot):
         """
         return PageTemplate(rh=self, filename='index.mako').render()
 
+    def changeIndexer(self):
+        """
+        Render manage/changeIndexer page.
+
+        [Converted to VueRouter]
+        """
+        return PageTemplate(rh=self, filename='index.mako').render()
+
     def subtitleMissed(self, whichSubs=None):
         """
         Serve manageEpisodeStatus page.