+++ medusa/tv/series.py
@@ -131,6 +131,10 @@ class SaveSeriesException(Exception):
     """Generic exception used for adding a new series."""
 
 
+class ChangeIndexerException(Exception):
+    """Generic exception used for changing a shows indexer."""
+
+
 class SeriesIdentifier(Identifier):
     """Series identifier with indexer and indexer id."""
 
@@ -1377,8 +1381,15 @@ class Series(TV):
         season_all_poster_result = metadata_provider.create_season_all_poster(self) or season_all_poster_result
         season_all_banner_result = metadata_provider.create_season_all_banner(self) or season_all_banner_result
 
-        return (fanart_result or poster_result or banner_result or season_posters_result or
-                season_banners_result or season_all_poster_result or season_all_banner_result)
+        return (
+            fanart_result
+            or poster_result
+            or banner_result
+            or season_posters_result
+            or season_banners_result
+            or season_all_poster_result
+            or season_all_banner_result
+        )
 
     def make_ep_from_file(self, filepath):
         """Make a TVEpisode object from a media file.
@@ -2306,6 +2317,7 @@ class Series(TV):
         data['id']['imdb'] = self.imdb_id
         data['id']['slug'] = self.identifier.slug
         data['id']['trakt'] = self.externals.get('trakt_id')
+        data['externals'] = {k.split('_')[0]: v for k, v in self.externals.items()}
         data['title'] = self.title  # Name plus (optional) year.
         data['name'] = self.name
         data['indexer'] = self.indexer_name  # e.g. tvdb